Update accessibility statement
We previously said we'd be able to fix the issues in the Tech Docs
Template (which affect the content API docs and the GOV.UK developer
docs among others) by the end of 2020. But we're now aiming for end of
March 2021.

Trello card: https://trello.com/c/c90no9QJ/2269-make-amendments-to-the-accessibility-statements-on-our-documentation