Remove duplicate link in header
There's already a link to `/` in the header (The main "GOV.UK Content API" link).

Having the same link in the header twice was flagged as an accessibility issue in [a recent audit of the tech docs template](https://docs.google.com/spreadsheets/d/1s53eGMBj2hdTVLRitwi673R-_uERWh0kOxOp9QsOVg0/edit#gid=0). This can't be fixed in the gem, but we can fix it in our docs.