Remove use of `widdershins`
We are using an old fork of the `widdershins` package, which has not
been updated for many years.

The need for using this package has been removed, since the
`tech-docs-gem` now has the ability to create API reference pages.

Therefore we can completely remove `widdershins` and the use of node
packages in this application.