Exclude JavaScript from CodeQL analysis
CodeQL analysis is currently failing (on CodeQL 2.32.5) due to finding
JavaScript files but not analysable JavaScript code. A workaround for
this is to tell CodeQL not to analyse the languages for which there's no
analysable code. The govuk-infrastructure action that wraps around the
CodeQL action has been updated to accept a `languages` input value which
will be passed down to the CodeQL action

actions (GitHub Actions) and ruby are the two other auto-detected
languages

https://github.com/alphagov/govuk-infrastructure/pull/3212

If analysable code in JavaScript or another supported language is
subsequently introduced, the specified languages will need updating (not
ideal)

https://docs.github.com/en/code-security/code-scanning/creating-an-advanced-setup-for-code-scanning/customizing-your-advanced-setup-for-code-scanning#changing-the-languages-that-are-analyzed